mvp game page (#104)
* mvp game page

* life is pain and we live to suffer. this was supposed to be fun, but
it's 2 am and i'm writing some of the worst code i've ever written. i'm
not proud of this, there is a universe where i turn this already long
night longer and fix this mess, but that is not his universe. the
war crimes committed will remain. I doubt anyone will care enough to fix
them, heck, no one cared enough to make sure they never occured in the
first place. but somehow I still care enough for this to sting.

* to fix build

* fixed minor bug with leaderboard time display

* hide category display once game is finished

* ocd

* submit button

* added word requests

* hardcoded seed

* hard code db url for demo

* css fix

* the fixed that would've saved the demo 😭OC

* fixed errors